Калькулятор
Сделал возвращаемые типы long, сделал кастинг в лямбда выражениях

+++ src/main/java/faang/school/godbless/calculator/Calculator.java
@@ -2,5 +2,5 @@ package faang.school.godbless.calculator;
 
 @FunctionalInterface
 public interface Calculator {
-    Integer calculate(Integer firstInteger, Integer secondInteger);
+    Long calculate(Integer firstInteger, Integer secondInteger);
 }
